verkle: port over some tests to util
