+++ packages/util/test/verkle.spec.ts
@@ -1,16 +1,21 @@
-import {
-  Address,
-  type VerkleCrypto,
-  type VerkleExecutionWitness,
-  bytesToHex,
-  hexToBytes,
-  randomBytes,
-} from '@ethereumjs/util'
 import { loadVerkleCrypto } from 'verkle-cryptography-wasm'
 import { assert, beforeAll, describe, it } from 'vitest'
 
 import * as verkleBlockJSON from '../../statemanager/test/testdata/verkleKaustinen6Block72.json'
-import { getVerkleStem, verifyProof } from '../src/index.js'
+import {
+  Address,
+  type VerkleCrypto,
+  type VerkleExecutionWitness,
+  VerkleLeafType,
+  bytesToHex,
+  concatBytes,
+  getVerkleKey,
+  getVerkleStem,
+  hexToBytes,
+  intToBytes,
+  randomBytes,
+  verifyVerkleProof,
+} from '../src/index.js'
 
 describe('Verkle cryptographic helpers', () => {
   let verkle: VerkleCrypto
@@ -36,20 +41,37 @@ describe('Verkle cryptographic helpers', () => {
     )
   })
 
-  it('verifyProof(): should verify verkle proofs', () => {
+  it('verifyVerkleProof(): should verify verkle proofs', () => {
     // Src: Kaustinen6 testnet, block 71 state root (parent of block 72)
     const prestateRoot = hexToBytes(
       '0x64e1a647f42e5c2e3c434531ccf529e1b3e93363a40db9fc8eec81f492123510'
     )
     const executionWitness = verkleBlockJSON.executionWitness as VerkleExecutionWitness
-    assert.isTrue(verifyProof(verkle, prestateRoot, executionWitness))
+    assert.isTrue(verifyVerkleProof(verkle, prestateRoot, executionWitness))
   })
 
-  it('verifyProof(): should return false for invalid verkle proofs', () => {
+  it('verifyVerkleProof(): should return false for invalid verkle proofs', () => {
     // Random preStateRoot
     const prestateRoot = randomBytes(32)
     const executionWitness = verkleBlockJSON.executionWitness as VerkleExecutionWitness
     // Modify the proof to make it invalid
-    assert.isFalse(verifyProof(verkle, prestateRoot, executionWitness))
+    assert.isFalse(verifyVerkleProof(verkle, prestateRoot, executionWitness))
+  })
+})
+
+describe('should generate valid tree keys', () => {
+  it('should generate valid keys for each VerkleLeafType', () => {
+    const stem = hexToBytes('0x318dea512b6f3237a2d4763cf49bf26de3b617fb0cabe38a97807a5549df4d')
+    for (const leaf of [
+      VerkleLeafType.Version,
+      VerkleLeafType.Balance,
+      VerkleLeafType.Nonce,
+      VerkleLeafType.CodeHash,
+      VerkleLeafType.CodeSize,
+    ]) {
+      const key = getVerkleKey(stem, leaf)
+      assert.equal(key.length, 32)
+      assert.deepEqual(key, concatBytes(stem, intToBytes(leaf)))
+    }
   })
 })
